refactor: eventcomment 도메인 코드 리팩토링
